Move 'datastore.exceptions_catch_remap_gax_error' to 'core.exceptions'.
Toward #3175.
